requirements.txt
robotframework>=5.0.1
robotframework-jsonlibrary>=0.5
robotframework-pabot>=2.15.0
robotframework-seleniumlibrary>=6.0.0
robotframework-zoomba>=3.4.4